Update patch set 3
Patch Set 3:

Understood Ying. It is based on what we have had in out dalvik-dev branch, although that was art only, so I had to refactor things to support switching between Dalvik and ART for now. I did test user builds with both Dalvik and ART as well as userdebug and eng for Dalvik (all this in internal master) then retested in AOSP master.

Patch-set: 3
Label: Verified=0