Update patch set 2
Patch Set 2:

These CFLAGS should not have been here in the first place. These are rules for controlling the linker, not C/C++ compilation. As far as the other updates to TARGET_GLOBAL_CFLAGS, I believe that is fine, since those are all cross-compiler flags that are accepted by both gcc and clang. I would prefer to get this change in now, as it is holding up an LLVM update internally. Any further fixes for these flags should be handled by Intel.

Patch-set: 2