Remove external/chromium from the 64 bit blacklist.
This isn't the same as chromium_org (which is used by the
webview). This is a 3 year old snapshot of the net stack which
compiles under 64 bit. It's currently used on aosp master by
libstagefright_chromium_http.

This project can be removed if the changes that remove the
chromium dependency are cherry-picked to aosp/master.

Change-Id: I5d0f9ed03ea9842b47d980d77ea32bc7a3f6998f