improve density selection of recovery assets
Some recovery UI assets are available in different sizes; we need to
pick the right ones for the device's density.  Use
PRODUCT_AAPT_PREF_CONFIG if it is available (it is empty for older
devices), and fall back to the highest density in PRODUCT_AAPT_CONFIG
if the PREF_ version is not available.

Change-Id: Ia1eb7d5409cda17b0fc5c0bcfc33dfa6a50893e8