Add libsigchain to the embedded.mk makefile
This library is used in a LD_PRELOAD environment variable and the linker shows
a warning if it is not present.  This change is to added it to all targets so
that the warning is not shown.  If ART is not provided on the device this
library has no effect.

Change-Id: I81f01b66fa9c78048534f4bcfc04d62b18057a03