Update patch set 11
Patch Set 11:

this is good. it lets me do:

  adb shell setprop debug.db.uid 20000
  adb shell crasher thread-abort

  gdbclient /system/xbin/crasher :5039 crasher

which wouldn't work before.

in a future change i think we can do a little better. as you demonstrate in this patch, we can get the path to the process' executable from /proc. so i should really be able to just say "gdbclient crasher :5039" and have it work out the /system/xbin/crasher bit for itself.

Patch-set: 11