envsetup.sh: Fix handling of emulator prebuilts path.
Due to a bad previous merge, everytime 'lunch' is called, a new
instance of prebuilts/android-emulator/<system>/ will be added
to your PATH.

This patch fixes the issue by removing the duplicate code that
should have been removed during the merge. Note that android-emulator
is still being added once to the PATH by code that appears before
in the script.

BUG=17524154

Change-Id: Ic281f71ad640e52d744f968f5c02b428cf6310dc