envsetup.sh: Fix handling of emulator prebuilts path.
Due to a bad previous merge, everytime 'lunch' is called, a new
instance of prebuilts/android-emulator/<system>/ will be added
to your PATH.

This patch fixes the issue by removing the duplicate code that
should have been removed during the merge. Note that android-emulator
is still being added once to the PATH by code that appears before
in the script.

BUG=17524154

Change-Id: Iaa03159b0c81b15ef4c7c420e2f76e0a48aba789