core: Let the build specify mount options for recovery
Currently recovery install/patch will mount /system++ with only default
options.
The default options are not very suitable for dealing with crashes.
We now use
 TARGET_RECOVERY_FSTYPE_MOUNT_OPTIONS
that can specify which mount options to use for a given fs-type.
  # <fstype>=<fstype_opts>[|<fstype_opts>]...
  # fstype_opts := <opt>[,<opt>]...
  #         opt := <name>[=<value>]
If only defaults are to be used, then the BoardConfig.mk can just have
an empty (must have ""):
 TARGET_RECOVERY_FSTYPE_MOUNT_OPTIONS := ""

The default is
 "ext4=match_batch_time=0,commit=1,data=journal,barrier=1,errors=panic,nodelalloc"

The release tools will pick them up with
  https://googleplex-android-review.git.corp.google.com/#/c/573628

Bug: 18092222
Change-Id: Iaa59099b0edf5af8835d1862fada089cff58882c