Remove logic for platform stlport gtest.
Now that the platforms stlport linked gtest is gone, we don't need the
logic for this any more. All platform tests will use libc++ (which has
been renamed to libgtest without a suffix).

Note that tests built with the NDK will still use stlport, but they
will use the NDK's stlport, not the platform's.

Bug: 15193147
Change-Id: Ifcb2489ed60c41e4ed028096b516c1ee643de907