Remove ARCH_ARM_HAVE_TLS_REGISTER.
This was only needed for platform builds, and was removed in the gap between
armv5 being removed and being added back again for tapas builds. Otherwise it
would have been removed back then.

(I suspect that we don't need the system.prop file in here, either, but I
don't really know how to test that.)

Change-Id: I212ff7b3568b5d5ff3cc66150ec7c4fa0b8cac92