Make sure DEX2OAT_TARGET_CPU_VARIANT is not empty.
On non-arm architectures, there is no TARGET_CPU_VARIANT set. For x86,
art assumes that the CPU variant is actually the TARGET_ARCH_VARIANT.
Therefore, if no TARGET_CPU_VARIANT is set, use the TARGET_ARCH_VARIANT.
If TARGET_ARCH_VARIANT is not set, then use default as the value.

Change-Id: I17fc9e3ac7412462103b8f0b287fce106298b741