Update patch set 2
Patch Set 2:

(2 comments)

The main function correctly returns 1 when it finds non-whitelisted classes. However the problem is that 'return 1' does not mean 'exit code is 1'. Please recall that the exit code of a python program is determined by the argument given to sys.exit(), not by the return value of the main function.

Patch-set: 2