Update patch set 1
Patch Set 1: Code-Review-1

libselinux will show up in others ... I grouped libcutils liblog and libselinux as a requirement for the fs_pstore patch sets (the bug this is tagged to).

zipalign is the first DISTTOOL that will start using 2 of the three shared libraries; with approval here, I can work through the other DISTOOLs that will use 1, 2 or 3 of these libraries. Because zipalign is also part of the SDK package, it is _not_ a good candidate example for switching over the DISTTOOL to using shared libraries.

SDK build tools package ... <sigh> ... yet another roadblock on the way. For this one, liblog and libcutils are (or should be) available to build for all host platforms.

So, basically, the change for disttools is ok, but this one requires far too much structural work and risk to be considered to help unblock fs_config additions in libcutils.

Patch-set: 1
Label: Code-Review=-1