Error out if LOCAL_PREBUILT_JNI_LIBS and LOCAL_CERTIFICATE := PRESIGNED are used
LOCAL_PREBUILT_JNI_LIBS is an indication to the build system that
all shared libraries should be deleted from an APK, and the shared
libraries should be placed in the application's /system/app directory.

However, using this option isn't appropriate for pre-signed APKs.
Any attempt to delete files from a pre-signed APK will corrupt it's
signature or waste disk space.

Bug: 20247329
Bug: 8076853
Bug: 1162500
Change-Id: I89ce8f06d3889dd79dd9ffe86fc5fa60814498ad