Update patch set 4
Patch Set 4:

> Verified checkbuild of this change + https://android-review.googlesource.com/#/c/148336/3
 >  (should be landed simultaneously).

fyi: If you set the same topic of both of them (with the little button to the right of the topic field above cherry-pick, rebase, etc) gerrit will enforce this and allow submitting both at the same time. You can also pass -t to repo upload to automatically set the topic to the local branch name.

Patch-set: 4