Remove LOCAL_MODULE from generated DBus header path.
Its presence requires #include directives to contain the build target
name, which is problematic because these directives can live in headers
that are shared by multiple build targets. Furthermore, having
LOCAL_MODULE in the generated header path is redundant because the
target directory is already private to the current build target (e.g.
.../<target_name>_intermediates/...).

Bug: 22608897
Change-Id: I059f71a1231e80f89c99441794a4491f2685036f