Update patch set 2
Patch Set 2:

Yes, that's the idea, to get rid of the whitelist and recursive make rule. There's currently a list of directories to load, and a separate list of modules to install in the windows sdk. Only the second list would stay around.

In the HOST_CROSS_OS change I'm already defaulting to not building windows modules unless they're explicitly required to be installed (only when win_sdk is a command line target, and no checkbuild/tags are used).

It may make sense to default to windows not supported, but I'll need to think about the syntax. I'll also need to check how many modules we'd need to whitelist that wouldn't otherwise require changes.

Patch-set: 2
Label: Verified=0