Update patch set 1
Patch Set 1: -Code-Review

Dan is right. And that seems to be the only reason to commit something with LOCAL_SANITIZE=thread to the repository ever - to force thread (or any other) sanitizer for a target.

We do not support 2 different sanitizers on device at the same time yet. We don't even have 2 different sanitizers on Android. When (and if) we do, we'll want LOCAL_SANITIZE=thread targets to build with thread sanitizer always.

Patch-set: 1