Update patch set 1
Patch Set 1:

I actually assumed the opposite. I assumed that if we had some test checked in with `LOCAL_SANITIZE := thread` but used `SANITIZE_TARGET=address` then we'd want to clobber the LOCAL_ in favor of our enviornment.

The reason for this being that while the test owner intends for it to be used to find thread issues, when we SANITIZE_TARGET=address is means we're interested in asan failures, not tsan failures.

Patch-set: 1