Update patch set 1
Patch Set 1:

asan_test is part of the ASan test suite. It will only pass if built with ASan. It's not part of the user image, of course.

It looks like LOCAL_SANITIZE is actually two options in one:
- "never" is used to exclude targets from SANITIZE_HOST/SANITIZE_TARGET
- address/thread/integer/etc is used to apply sanitization by default.

SANITIZE_TARGET=address means "build as much as possible with ASan".
From this point of view, it should merge with compatible sanitizers, like integer, and override incompatible ones, like thread.

To define a target that works when built with, say, "thread", we should do something special, ex.

if (SANITIZE_TARGET empty or "thread")
LOCAL_MODULE:=..
LOCAL_SANITIZE:=thread
...
endif

Patch-set: 1