DO NOT MERGE releasetools: Always write the last block if it's padded.
In BBOTAs if the last block of a DataImage is padded, we should always
write the whole block even for incremental OTAs. Because otherwise the
last block may be skipped if unchanged, but would fail the post-install
verification if it has non-zero contents in the padding bytes.

Bug: 24981812
Change-Id: Id0608384fcff0375aec8c158e6f909569b873d31
(cherry picked from commit 7589e961a7b8b7d964704db87f83ed8811e97ac6)