emulator.mk: Add ethernet permission
Networking on old emulator device(a.k.a. goldfish) relies on RIL;
However, RIL is not available on ranchu device yet. For ranchu device
to have a functional networking, we need to add ethernet permission
and start dhcpcd_eth0 service on ranchu device. This CL adds ethernet
permission to both goldfish (which is harmless) and ranchu devices.

This addition was originally proposed by miroslav.tisma@imgtec.com
and the ethernet permission was only added to arm64 and mips64 after
discussion. With the ranchu device supporting more architectures,
it makes sense to allow this permission on all devices and all CPU
architectures.

related CL by miroslav.tisma@imgtec.com on AOSP:
    b09fb84becf73bf2bc578ebf27910d75b79d668a

bug: 24070972

Change-Id: I81ac5d8901adee43784fe9dd45a170fb90bb824a