Pack META/ab_partitions.txt when using the A/B OTA updater.
The list of A/B partitions updated is product-specific. Normally you
would update boot and system, but the product could require to update
other partitions during a system update. This patch packages the list
of A/B partitions configured to be updated when using the A/B updater.
The list can be set in the product makefile by setting the make
variable UPDATE_AB_PARTITIONS.

Bug: 24387863
Test: `make dist` creates the META/ab_partitions.txt in the target_files.zip

Change-Id: I4226ab05c7c7cd74721b06f1da64c6cf4d0e52ac